feat: 完成问题组的逻辑

+++ cherino/crud/auth.py
@@ -2,7 +2,13 @@ from typing import Optional
 
 from peewee import fn
 
-from cherino.database.models import Answer, AnswerHistory, PendingVerify, Question
+from cherino.database.models import (
+    Answer,
+    AnswerHistory,
+    PendingVerify,
+    Question,
+    GroupQuestion,
+)
 
 
 def add_question(
@@ -15,6 +21,10 @@ def add_question(
     question = Question.create(
         group=chat_id, description=description, image=image, correct_answer=0
     )
+    GroupQuestion.insert(
+        group=chat_id, questions=question
+    ).on_conflict_ignore().execute()
+
     correct_answer, *other_answers = answers
     for answer in other_answers:
         Answer.insert(question=question, description=answer).execute()
@@ -27,17 +37,22 @@ def get_all_questions(chat_id: int) -> list[Question]:
     """
     获取所有入群问题
     """
-    return list(Question.select().where(Question.group == chat_id))
+    return list(
+        Question.select()
+        .join(GroupQuestion, on=(Question.group == GroupQuestion.questions))
+        .where(GroupQuestion.group == chat_id)
+    )
 
 
 def delete_question(chat_id: int, question_id: int):
     """
     删除一个问题
     """
+    question = Question.get_by_id(question_id)
+    if question.group != chat_id:
+        return
     Answer.delete().where(Answer.question == question_id).execute()
-    Question.delete().where(
-        Question.id == question_id, Question.group == chat_id
-    ).execute()
+    question.delete_instance()
 
 
 def get_question(chat_id: int) -> Optional[tuple[Question, Answer]]:
@@ -47,7 +62,8 @@ def get_question(chat_id: int) -> Optional[tuple[Question, Answer]]:
     try:
         question = (
             Question.select()
-            .where(Question.group == chat_id)
+            .join(GroupQuestion, on=(Question.group == GroupQuestion.questions))
+            .where(GroupQuestion.group == chat_id)
             .order_by(fn.Random())
             .limit(1)
             .get()
@@ -93,3 +109,37 @@ def get_pending_verify(
         )
     else:
         return list(PendingVerify.select().where(PendingVerify.group == chat_id))
+
+
+def add_group_question(chat_id: int, group_id: int):
+    """
+    添加一个来自其他群组的题库
+    """
+    GroupQuestion.insert(
+        group=chat_id, questions=group_id
+    ).on_conflict_ignore().execute()
+
+
+def answer_stats(chat_id: int) -> list[tuple[Question, int, int]]:
+    """
+    查询当前群组题库的回答状态，返回一个列表，每个元素为一个三元组：题目、总回答、正确回答
+    """
+    query = (
+        AnswerHistory.select(
+            AnswerHistory.question,
+            fn.COUNT(AnswerHistory.question),
+            fn.SUM(
+                fn.CASE(
+                    None,
+                    ((AnswerHistory.answer == Question.correct_answer, 1),),
+                    0,
+                )
+            ),
+        )
+        .join(Question, on=(AnswerHistory.question == Question.id))
+        .where(Question.group == chat_id)
+        .group_by(AnswerHistory.question)
+    )
+    return [
+        (Question.get_by_id(q), total, correct) for q, total, correct in query.tuples()
+    ]
